+++ app/components/Views/confirmations/components/info/switch-account-type/switch-account-type.tsx
@@ -4,7 +4,7 @@ import { View } from 'react-native';
 import { useConfirmationMetricEvents } from '../../../hooks/metrics/useConfirmationMetricEvents';
 import AccountNetworkInfo from '../../rows/account-network-info-row';
 import AdvancedDetailsRow from '../../rows/transactions/advanced-details-row/advanced-details-row';
-import GasFeesDetails from '../../rows/transactions/gas-fee-details';
+import GasFeesDetailsRow from '../../rows/transactions/gas-fee-details-row';
 import SwitchAccountTypeInfoRow from '../../rows/switch-account-type-info-row';
 
 const SwitchAccountType = () => {
@@ -15,7 +15,7 @@ const SwitchAccountType = () => {
     <View>
       <AccountNetworkInfo />
       <SwitchAccountTypeInfoRow />
-      <GasFeesDetails />
+      <GasFeesDetailsRow />
       <AdvancedDetailsRow />
     </View>
   );
